backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pdfplumber==0.10.3
sentence-transformers==2.2.2
chromadb==0.4.17
google-generativeai==0.3.0
python-dotenv==1.0.0
requests==2.31.0
cors==1.0.1
